Atualizando ServiceSolicitacao

+++ src/main/java/br/com/ifba/solicitacao/service/ServiceSolicitacao.java
@@ -4,6 +4,7 @@
  */
 package br.com.ifba.solicitacao.service;
 
+import br.com.ifba.infrastructure.exception.BusinessException;
 import br.com.ifba.solicitacao.dao.IDaoSolicitacao;
 import br.com.ifba.solicitacao.model.Solicitacao;
 import java.util.List;
@@ -16,11 +17,27 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class ServiceSolicitacao implements IServiceSolicitacao {
+   //================= CONSTANTES =============================================
+    
+   // Mensagem de erro se a Solicitacao for null.
+    public final static String SOLICITACAO_NULL = "Dados de Solicitacao nao preenchidos";
+    
+    // Mensagem de erro se a Solicitacao ja existir no banco.
+    public final static String SOLICITACAO_EXISTE = "Solicitacao ja existente no Banco de dados";
+   
+    // Mensagem de erro se a Solicitacao não existir no banco.
+    public final static String SOLICITACAO_NAO_EXISTE = "Solicitacao nao existente no Banco de dados";
+    
     @Autowired
     private IDaoSolicitacao daoSolicitacao;
     
     @Override
     public void saveSolicitacao(Solicitacao solicitacao) {
+        if(solicitacao == null)
+            throw new BusinessException(SOLICITACAO_NULL);
+        if(daoSolicitacao.existsById(solicitacao.getId())) 
+            throw new BusinessException(SOLICITACAO_EXISTE);
+
         daoSolicitacao.save(solicitacao);
     }
     
@@ -31,6 +48,9 @@ public class ServiceSolicitacao implements IServiceSolicitacao {
     
     @Override
     public void removeSolicitacao(Long value){
+        if(!daoSolicitacao.existsById(value)) 
+            throw new BusinessException(SOLICITACAO_NAO_EXISTE);
+
         daoSolicitacao.deleteById(value);
     }
 }
